Initial commit. Completed all in youtube video Part 1
In the README.md I link two videos from the same Youtube video series
which is what I closely follow when doing this project, with this
commit I have accomplished everything that is covered in the
first out of these two videos.

Additional to writing the same code he does in his video,
I have added quite a lot of information in code-comments
and README.md file as footnotes to understand everything
that goes on and why we do each step etc.

So far I've primarily documented in the following places:
- README.md
- Makefile
- targets/x86_64/linker.ld
- buildenv/Dockerfile
- src/impl/x86_64/boot/header.asm
- src/impl/x86_64/boot/main.asm

I think I can still document more in the
following files for future commits:
- targets/x86_64/linked.ld
- targets/x86_64/iso/boot/grub/grub.cfg
- src/impl/x86_64/boot/header.asm
- src/impl/x86_64/boot/main.asm